# Add fast Unsigned->Word conversions