--- a/clash-lib/prims/common/Clash_Sized_Internal_Unsigned.json
+++ b/clash-lib/prims/common/Clash_Sized_Internal_Unsigned.json
@@ -19,4 +19,28 @@
     , "template"  : "~ARG[0] / ~ARG[1]"
     }
   }
+, { "BlackBox" :
+    { "name"      : "Clash.Sized.Internal.Unsigned.unsignedToWord"
+    , "kind"      : "Expression"
+    , "template"  : "~ARG[0]"
+    }
+  }
+, { "BlackBox" :
+    { "name"      : "Clash.Sized.Internal.Unsigned.unsigned8toWord8"
+    , "kind"      : "Expression"
+    , "template"  : "~ARG[0]"
+    }
+  }
+, { "BlackBox" :
+    { "name"      : "Clash.Sized.Internal.Unsigned.unsigned16toWord16"
+    , "kind"      : "Expression"
+    , "template"  : "~ARG[0]"
+    }
+  }
+, { "BlackBox" :
+    { "name"      : "Clash.Sized.Internal.Unsigned.unsigned32toWord32"
+    , "kind"      : "Expression"
+    , "template"  : "~ARG[0]"
+    }
+  }
 ]