Usefull when Clash simulation needs to interact with the outside world; i.e. for `Unsigned 8 -> Ptr Word8`

before:
```
benchmarking Unsigned/unsigned8toWord8 WORD_SIZE_IN_BITS
time                 21.89 ns   (21.77 ns .. 22.05 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 21.96 ns   (21.81 ns .. 22.38 ns)
std dev              806.7 ps   (185.9 ps .. 1.658 ns)
variance introduced by outliers: 59% (severely inflated)

benchmarking Unsigned/unsigned16toWord16 WORD_SIZE_IN_BITS
time                 22.48 ns   (22.30 ns .. 22.77 ns)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 22.54 ns   (22.37 ns .. 23.04 ns)
std dev              914.9 ps   (351.0 ps .. 1.801 ns)
variance introduced by outliers: 64% (severely inflated)

benchmarking Unsigned/unsigned32toWord32 WORD_SIZE_IN_BITS
time                 22.45 ns   (22.44 ns .. 22.47 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 22.48 ns   (22.45 ns .. 22.56 ns)
std dev              139.2 ps   (43.39 ps .. 279.7 ps)

benchmarking Unsigned/unsignedToWord WORD_SIZE_IN_BITS
time                 21.93 ns   (21.87 ns .. 22.03 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 21.89 ns   (21.86 ns .. 21.96 ns)
std dev              143.9 ps   (42.29 ps .. 275.4 ps)
```

after:
```
benchmarking Unsigned/unsigned8toWord8 WORD_SIZE_IN_BITS
time                 4.559 ns   (4.558 ns .. 4.562 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 4.565 ns   (4.561 ns .. 4.573 ns)
std dev              17.88 ps   (10.75 ps .. 32.04 ps)

benchmarking Unsigned/unsigned16toWord16 WORD_SIZE_IN_BITS
time                 4.783 ns   (4.781 ns .. 4.785 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 4.783 ns   (4.782 ns .. 4.784 ns)
std dev              4.618 ps   (3.615 ps .. 6.532 ps)

benchmarking Unsigned/unsigned32toWord32 WORD_SIZE_IN_BITS
time                 5.086 ns   (5.085 ns .. 5.088 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 5.087 ns   (5.085 ns .. 5.091 ns)
std dev              7.924 ps   (4.884 ps .. 14.50 ps)

benchmarking Unsigned/unsignedToWord WORD_SIZE_IN_BITS
time                 4.667 ns   (4.664 ns .. 4.671 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 4.675 ns   (4.666 ns .. 4.713 ns)
std dev              57.20 ps   (6.486 ps .. 120.4 ps)
variance introduced by outliers: 15% (moderately inflated)
```